Only updating health text when health changes, changed health to int, removed a lot of allocations

+++ gameSrc/bbth/game/units/DefendingUnit.java
@@ -134,7 +134,7 @@ public class DefendingUnit extends Unit {
 	}
 
 	@Override
-	public float getStartingHealth() {
+	public int getStartingHealth() {
 		return 60;
 	}
